pyproject.toml
[tool.poetry]
name = "kafka-elasticsearch-analytics"
version = "0.1.0"
description = "Real-Time E-commerce Analytics & Search Platform with Kafka and Elasticsearch"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
faker = "^20.0.0"
kafka-python = "^2.0.2"
elasticsearch = "^8.11.0"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
pydantic = "^2.5.0"
python-dotenv = "^1.0.0"
httpx = "^0.25.2"
pandas = "^2.1.3"
numpy = "^1.26.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.11.0"
ruff = "^0.1.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

